+++ src/pages/Processos/ImportacaoDados/Habilitados/components/Historico/index.tsx
@@ -3,9 +3,10 @@ import { DeleteOutlined } from "@ant-design/icons";
 import type { ColumnsType, TableProps } from "antd/es/table";
 import dayjs from "dayjs";
 import { Button, Space, Typography } from "antd";
-import { StyledTable } from "../../../../../../components/estilosCompartilhados/styles";
+import { SecondaryButton,StyledTable } from "../../../../../../components/estilosCompartilhados/styles";
 import { useImportacaoDados } from "../../hooks/useImportacaoDados";
-import { SecondaryButton } from "../../styles";
+
+
 
 const { Title } = Typography;
