+++ src/pages/Processos/ImportacaoDados/Habilitados/index.tsx
@@ -15,7 +15,7 @@ import {
   ActionButtonsContainer,
   SecondaryButton,
   PrimaryButton,
-} from "./styles";
+} from "../../../../components/estilosCompartilhados/styles";
 import { useConcursos } from "../../NovaConvocacaoCandidatos/hooks/useConcursos";
 
 
@@ -93,7 +93,7 @@ const Habilitados: React.FC<HabilitadosProps> = ({ onShowHistorico, onShowLayout
                   help={formErrors.arquivo?.message}
                   labelCol={{ span: 24 }}
                 >
-                  <StyledUpload                  
+                  <StyledUpload
                     beforeUpload={(file) => {
                       handleFileUpload(file);
                       return false; // Impede o upload automático
@@ -144,3 +144,5 @@ const Habilitados: React.FC<HabilitadosProps> = ({ onShowHistorico, onShowLayout
 };
 
 export default Habilitados;
+
+
