+++ src/pages/Processos/ImportacaoDados/Vagas/hooks/useImportacaoVagasSchema.tsx
@@ -2,11 +2,12 @@ import * as yup from "yup";
 
 const useImportacaoVagasSchema = () => {
   return yup.object({
+    ignorar_primeira_linha:yup.boolean(),
     metodo_de_importacao: yup.number().required("Método de importação é obrigatório"),
 
     cargo: yup.string().when("metodo_de_importacao", {
       is: 1,
-      then: (schema) => schema.required("Cargo é obrigatório"),
+      then: (schema) => schema.required("Cargo é obrigatório, selecione um cargo"),
       otherwise: (schema) => schema.notRequired(),
     }),
